requirements.txt
fastapi
uvicorn[standard]
httpx
pillow
opencv-python
ultralytics
numpy
